- Re-Indexed Curse games with ids - Moved Environment.ts to util/ - Moved embedBuilder.ts to discord/
