+++ src/util/discord.ts
@@ -1,6 +1,6 @@
-import { Snowflake } from 'discord-api-types/globals';
-import { ActivityType, Client, ClientUser, PermissionFlagsBits, PermissionsBitField } from 'discord.js';
-import Environment from '../model/Environment';
+import {Snowflake} from 'discord-api-types/globals';
+import {ActivityType, Client, ClientUser, PermissionFlagsBits, PermissionsBitField} from 'discord.js';
+import Environment from './Environment';
 
 ///The different levels of permission that may be needed to execute a certain command
 export enum CommandPermission {
